Cleanup / Code-Style
